requirements.txt
beautifulsoup4
requests
selenium
pandas
scikit-learn
matplotlib
seaborn
spacy
sentence-transformers
torch
pymongo
psycopg2-binary
elasticsearch